Revert "build(rust): split vendored C API source"
This reverts commit 87eecebdd65d0de648e6b647ac05dc949d4e0096.
